9.15.24 cumulative
Finished Stripe checkout
Can add and remove cart products
Updated home page and hero
Created Orders page to track all orders; Checks paid/unpaid
Created Order page by [id]; Landing page after Stripe
Added flying-cart-item effect
Navbar responsive; Mobile popup window
Added function to determine if user is admin
Updated routes; Must be admin for certain functions

Fixed multiple bugs & errors
Made all pages responsive; Updated images with CC0 versions

+++ src/app/api/categories/route.js
@@ -1,31 +1,43 @@
 import { Category } from "@/models/category"
 import mongoose from "mongoose"
+import { isAdmin } from "../auth/[...nextauth]/route"
+
+export async function GET() {
+    mongoose.connect(process.env.MONGO_URL)
+    return Response.json(
+        await Category.find() // Find all categories
+    )
+}
 
 export async function POST(req) {
     mongoose.connect(process.env.MONGO_URL)
     const { name } = await req.json()
-    const categoryDoc = await Category.create({name})
-    return Response.json(categoryDoc)
+    if (await isAdmin()) {
+        const categoryDoc = await Category.create({name})
+        return Response.json(categoryDoc)
+    } else {
+        return Response.json({})
+    }
 }
 
 export async function PUT(req) {
     mongoose.connect(process.env.MONGO_URL)
     const {_id, name} = await req.json()
-    await Category.updateOne({_id}, {name}) // Update document with given id; Change name
+    if (await isAdmin()) {
+        await Category.updateOne({_id}, {name}) // Update document with this id; Change name
+    }    
     return Response.json(true)
 }
 
-export async function GET() {
-    mongoose.connect(process.env.MONGO_URL)
-    return Response.json(
-        await Category.find() // Use find to get all categories
-    )
-}
+
 
 export async function DELETE(req) {
     mongoose.connect(process.env.MONGO_URL)
     const url = new URL(req.url)
     const _id = url.searchParams.get('_id')
-    await Category.deleteOne({_id})
+
+    if(await isAdmin()) {
+        await Category.deleteOne({_id}) // Delete document (category) with this id
+    }
     return Response.json(true)
 }
